% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_read.R
\name{csv_read}
\alias{csv_read}
\title{Sample csv data set for analysis}
\usage{
csv_read(path)
}
\arguments{
\item{path}{path to filename}
}
\value{
a \code{tibble}
}
\description{
Sample csv data set for analysis
}
\examples{
csv = system.file("extdata", "csv_data.csv", package = "hglmbc2")
csv_read(csv)
}
