% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignMX.R
\name{DesignM}
\alias{DesignM}
\title{Creates design matrix for fixed effects}
\usage{
DesignM(data, fe.cont, fe.disc, ref.group)
}
\arguments{
\item{data}{The data frame.}

\item{fe.cont}{Vector of continuous variables.}

\item{fe.disc}{Vector of discrete/categorical variables.}

\item{ref.group}{Define the reference group for each discrete variable.}
}
\value{
A matrix.
}
\description{
\code{DesignM} function creates the design matrix for fixed effects.
}
\examples{
data <- eversmoke
re <- "county"
fe.disc <- c("year","gender","race","age")
fe.cont <- "povt_rate"
ref.group <- c("age1", "gender2","race4","year1")
}
