% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVars.R
\name{getVars}
\alias{getVars}
\title{Extract model variables from formula.}
\usage{
getVars(mformula = NULL)
}
\arguments{
\item{mformula}{the model formula.}
}
\value{
Returns a list with y, discrete variables, and continuous variables.
}
\description{
\code{myFormula} outputs the model formula based on discrete variables and continuous variables.
}
\seealso{
\code{\link{as.formula}}
}
