% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initPar.R
\name{initPar}
\alias{initPar}
\title{Obtain initial parameters}
\usage{
initPar(data, resp, dom, fe.disc = NULL, fe.cont = NULL, y.family = NULL, ...)
}
\arguments{
\item{data}{the data set.}

\item{resp}{the response variable.}

\item{dom}{domain.}

\item{fe.disc}{fixed effects (categorical)}

\item{fe.cont}{fixed effects (continuous)}

\item{y.family}{distribution of y given u}

\item{...}{other parameters.}
}
\value{

}
\description{
Obtain initial parameters
}
