% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFormula.R
\name{myFormula}
\alias{myFormula}
\title{model formula}
\usage{
myFormula(data, resp = NULL, dom = NULL, ...)
}
\arguments{
\item{data}{the data frame.}

\item{resp}{the response variable.}

\item{dom}{the random effect \code{dom} could be \code{"county","FIPS","state","zipcode"}.}

\item{...}{Other parameters. \code{fe.disc} A vector of discrete/categorical fixed effects. \code{fe.cont} A vector of continuous fixed effects.}
}
\value{
Returns an object of class for model \code{formula} which is a symbolic formula based on the variables in the data set.
}
\description{
\code{myFormula} outputs the model formula based on discrete variables and continuous variables.
}
\examples{
# Using discrete and continuous variables.
data <- eversmoke
y <- "smoke_ever"
dom <- "county"
catX <- c("age","gender","year","race")
contX <- "povt_rate"
\dontrun{
myFormula(data, resp = y, fe.disc = catX, fe.cont = contX)
as.formula(myFormula(data, resp = y, fe.disc = catX, fe.cont = contX))



# Only using discrete variables
resp <- "smoke_ever"
dom <- "county"
fe.disc <- c("age","gender","year")
myFormula(data=eversmoke, resp="smoke_ever", dom="county", fe.disc)
as.formula(myFormula(resp,fe.disc))



# None of the variables are defined, this only output the response variable.
resp <- "smoke_ever"
dom <- "county"
myFormula(data=eversmoke, resp="smoke_ever", dom="county")
}


}
\seealso{
\code{\link{as.formula}}
}
