% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predsFunc.R
\name{predsFunc}
\alias{predsFunc}
\title{Extract model variables from formula.}
\usage{
predsFunc(mformula = NULL)
}
\arguments{
\item{mformula}{the model formula.}
}
\value{
Returns a list with y, discrete variables, and continuous variables.
}
\description{
\code{myFormula} outputs the model formula based on discrete variables and continuous variables.
}
\examples{
# Using discrete and continuous variables.
mformula <- "smoke_ever ~ as.factor(age) + as.factor(gender) + as.factor(race)
+ as.factor(year) + povt_rate"
\dontrun{
predFunc(mformula)

}


}
\seealso{
\code{\link{as.formula}}
}
