% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_data.R
\docType{data}
\name{temp_data}
\alias{temp_data}
\title{Sample data set for testing functions}
\format{
A \code{data frame} with 3 columns, which are:
\describe{
\item{x1}{Normal random variable}
\item{x2}{Normal random variable with mean 4 and variance 2}
\item{cls}{Categorical variable with 5 groups}
\item{y}{Outcome that is created using x1 and \code{x2}}
}
}
\usage{
temp_data
}
\description{
A simple \code{data frame} that is for simple testing and runing exmaples.
}
\keyword{datasets}
