% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVars.R
\name{getVars}
\alias{getVars}
\title{Function to get discrete, numeric and response variables if only the model fit formula is defined.}
\usage{
getVars(mformula)
}
\arguments{
\item{mformula}{model formula}
}
\value{

}
\description{
Function to get discrete, numeric and response variables if only the model fit formula is defined.
}
